bump convex version
