+++ template-component/example/convex/_generated/server.d.ts
@@ -150,18 +150,4 @@ export type DatabaseReader = GenericDatabaseReader<DataModel>;
  */
 export type DatabaseWriter = GenericDatabaseWriter<DataModel>;
 
-export declare const components: {
-  counter: {
-    public: {
-      add: FunctionReference<
-        "mutation",
-        "internal",
-        { count: number; name: string; shards?: number },
-        null
-      >;
-      count: FunctionReference<"query", "internal", { name: string }, number>;
-    };
-  };
-};
-
 /* prettier-ignore-end */
