+++ template-component/src/component/_generated/api.d.ts
@@ -53,4 +53,6 @@ export declare const internal: FilterApi<
   FunctionReference<any, "internal">
 >;
 
+export declare const components: {};
+
 /* prettier-ignore-end */
